+++ cmd/runner_test.go
@@ -1,3 +1,16 @@
+//   Copyright © 2016, 2018, Oracle and/or its affiliates.  All rights reserved.
+//
+//   Licensed under the Apache License, Version 2.0 (the "License");
+//   you may not use this file except in compliance with the License.
+//   You may obtain a copy of the License at
+//
+//       http://www.apache.org/licenses/LICENSE-2.0
+//
+//   Unless required by applicable law or agreed to in writing, software
+//   distributed under the License is distributed on an "AS IS" BASIS,
+//   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+//   See the License for the specific language governing permissions and
+//   limitations under the License.
 package cmd
 
 import (
@@ -29,7 +42,7 @@ type MockStep struct {
 }
 
 // Mock methods not implemented by BaseStep
-func (s *MockStep) CollectArtifact(string) (*core.Artifact, error) {
+func (s *MockStep) CollectArtifact(context.Context, string) (*core.Artifact, error) {
 	return nil, nil
 }
 
@@ -63,11 +76,11 @@ type MockPipeline struct {
 }
 
 //Mock methods not implemented by BasePipeLine
-func (s *MockPipeline) CollectArtifact(string) (*core.Artifact, error) {
+func (s *MockPipeline) CollectArtifact(context.Context, string) (*core.Artifact, error) {
 	return nil, nil
 }
 
-func (s *MockPipeline) CollectCache(string) error {
+func (s *MockPipeline) CollectCache(context.Context, string) error {
 	return nil
 }
 
@@ -98,13 +111,14 @@ func (s *MockPipeline) Env() *util.Environment {
 //TestRunnerStepFailedOnInitEnvError tests the scenario when a step in the pipleine
 // will fail when an error occurs in initEnv() in step
 func (s *RunnerSuite) TestRunnerStepFailedOnInitEnvError() {
+	ctx := context.Background()
 	mockPipeline := &MockPipeline{}
 	shared := &RunnerShared{pipeline: mockPipeline}
 	step := &MockStep{}
 	runner := &Runner{}
 	runner.emitter = core.NewNormalizedEmitter()
 
-	sr, err := runner.RunStep(shared, step, 1)
+	sr, err := runner.RunStep(ctx, shared, step, 1)
 	s.Error(err)
 	fmt.Println(err.Error())
 	s.Contains(err.Error(), "Step initEnv failed with error message")