+++ docker/publish_step.go
@@ -1,4 +1,4 @@
-//   Copyright 2017 Wercker Holding BV
+//   Copyright © 2017,2018, Oracle and/or its affiliates.  All rights reserved.
 //
 //   Licensed under the Apache License, Version 2.0 (the "License");
 //   you may not use this file except in compliance with the License.
@@ -178,7 +178,7 @@ func (s *PublishStep) CollectFile(a, b, c string, dst io.Writer) error {
 }
 
 // CollectArtifact NOP
-func (s *PublishStep) CollectArtifact(string) (*core.Artifact, error) {
+func (s *PublishStep) CollectArtifact(context.Context, string) (*core.Artifact, error) {
 	return nil, nil
 }