+++ changelog.md
@@ -2,6 +2,7 @@
 
 - Changes for robust error handling and reporting in docker-push (#387)
 - Replace docker links with docker network (#382)
+- Change some docker API calls to use the official Docker client (#385) 
 
 ## v1.0.1201 (2018-04-16)